Combine source into single url_normalize module
